requirements/base.txt
# Core Django
Django==4.2.16
djangorestframework==3.15.2
django-cors-headers==4.4.0
django-filter==24.3
djangorestframework-simplejwt==5.3.1
django-extensions==3.2.3

# Database
psycopg2-binary==2.9.9
dj-database-url==2.2.0

# Cache / Queue
redis==5.1.1
django-redis==5.4.0
celery==5.4.0
kombu==5.3.7

# Storage
django-storages==1.14.4
Pillow==10.4.0

# NLP & ML
sentence-transformers==3.1.1
transformers==4.44.2
torch==2.4.1
scikit-learn==1.5.2
numpy==1.26.4
pandas==2.2.3
spacy==3.7.6

# Explainability
shap==0.46.0
lime==0.2.0.1

# Data generation
faker==30.3.0
kaggle==1.6.17
pdfminer.six==20231228
python-docx==1.1.2

# API clients
requests==2.32.3
httpx==0.27.2

# Utilities
python-decouple==3.8
python-dotenv==1.0.1
celery-progress==0.4
django-celery-results==2.5.1
django-celery-beat==2.7.0

# Serialization
djangorestframework-camel-case==1.4.2

# Logging
structlog==24.4.0
